refactor: dedupe history fetch into _gather_history helper
Collapses the identical context-fetch blocks in ask, code, and draw into
a single _gather_history(nick, channel, *, max_age_seconds=None) helper.
Also tightens max_age_seconds typing (int vs float) and drops a comment
that restated the commit message.

+++ plugins/llm/src/llm/context.py
@@ -224,7 +224,7 @@ class ConversationContext:
         channel: str,
         *,
         config: ContextConfig | None = None,
-        max_age_seconds: float | None = None,
+        max_age_seconds: int | None = None,
     ) -> list[dict[str, str]]:
         """Get conversation history for LiteLLM.
 
@@ -304,7 +304,7 @@ class ConversationContext:
         exclude_nick: str | None = None,
         *,
         config: ContextConfig | None = None,
-        max_age_seconds: float | None = None,
+        max_age_seconds: int | None = None,
     ) -> list[dict[str, str]]:
         """Get shared channel context.
